+++ pyBigKinds/__init__.py
@@ -32,4 +32,5 @@ __all__ = [
     "lsa",
     "kmeans",
     "dbscan",
+    "lda",
 ]
