+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyBigKinds
-version = 1.0.5
+version = 1.1.0
 description = BigKinds Data Analysis Toolkit for python
 author = Sorrychoe
 license = MIT
@@ -23,6 +23,7 @@ install_requires =
     pandas<2
     wordcloud>=1.8.2.2
     scikit-learn>=1.0.0
+    tomotopy>=0.12.4
     openpyxl>=3.1.2
 
 [options.extras_require]
